requirements.txt
biopython==1.79
numpy
scikit-learn
requests
python-dotenv
